requirements.txt
black==23.7.0
Jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0
ruff==0.0.278
sendgrid==6.10.0
